Move the example in tktable.py to examples/example1.py
